Implemented a custom higher quality transition for SaveToGalleryViewController
NOTE:
Using a proper transitionDelegate via SaveToGalleryTransitionManager FAILED. Not certain if my fault or iOS but I see a some reports on StackOverflow about a similar crash. I'm leaving the class in, because I may use it in the future. But it's not being used. Instead I implemented a faux transition by setting the modal overlay transition to crossfade and then in viewWillAppear I animate a springy plopping in of the dialog.

ALSO:
Fixed animation of dialog vertical offset when keyboard shows and hides by using an affine transform instead of animating the vertical centering constraint. Smoother!